Move entry to collection package
