+++ pkg/collection/entry/entry.go
