+++ pkg/collection/entry/metadata.go
